+++ merlin/celery.py
@@ -47,6 +47,7 @@ from merlin.config import (
     broker,
     results_backend,
 )
+from merlin.config.configfile import CONFIG
 from merlin.log_formatter import FORMATS
 from merlin.router import route_for_task
 
@@ -100,10 +101,10 @@ app.conf.update(
     redis_max_connections=100000,
 )
 
-# Set a 24-hour timeout to acknowledge a task before it's available to grab
-# again.
+# Set a timeout to acknowledge a task before it's available to grab
+# again (default 24 hours).
 app.conf.broker_transport_options = {
-    "visibility_timeout": 86400,
+    "visibility_timeout_seconds": CONFIG.celery.visibility_timeout_seconds,
     "max_connections": 100,
 }